Add a mechanism for changing project/language
On the ‘Thank you’ screen, when you click on the
TrueFootprint logo to restart the app, instead
show a dropdown that offers a choice of project
and language.

When selected, set the project/language state and
restart the app. The content is then sourced from
a static/content.js file which is passed into each
of the pages.